pyproject.toml
[project]
name = "epitag"
authors = [
    { name = "Applied Relevance", email = "geveritt@appliedrelevance.com"}
]
description = "Barcode and NFC printing and scanning support for ERPNext"
requires-python = ">=3.10"
readme = "README.md"
dynamic = ["version"]
dependencies = [
    # "frappe~=15.0.0" # Installed and managed by bench.
    "python-barcode~=0.15.1"
]

[build-system]
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

# These dependencies are only installed when developer mode is enabled
[tool.bench.dev-dependencies]
# package_name = "~=1.1.0"
